requirements.txt
eel
google-generativeai
gitignore-parser